Refactor package structure to separate network level configuration changes and snapshots from device level changes and snapshots.
