--- pkg/store/snapshot/store.go
@@ -1,147 +1,0 @@
-// Copyright 2019-present Open Networking Foundation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package snapshot
-
-import (
-	"context"
-	"github.com/atomix/atomix-go-client/pkg/client/map"
-	"github.com/atomix/atomix-go-client/pkg/client/primitive"
-	"github.com/atomix/atomix-go-client/pkg/client/session"
-	"github.com/atomix/atomix-go-local/pkg/atomix/local"
-	"github.com/atomix/atomix-go-node/pkg/atomix"
-	"github.com/atomix/atomix-go-node/pkg/atomix/registry"
-	"github.com/gogo/protobuf/proto"
-	"github.com/onosproject/onos-config/pkg/store/utils"
-	snapshottype "github.com/onosproject/onos-config/pkg/types/snapshot"
-	"github.com/onosproject/onos-topo/pkg/northbound/device"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/test/bufconn"
-	"io"
-	"net"
-	"time"
-)
-
-const snapshotsName = "snapshots"
-
-// NewAtomixStore returns a new persistent Store
-func NewAtomixStore() (Store, error) {
-	client, err := utils.GetAtomixClient()
-	if err != nil {
-		return nil, err
-	}
-
-	group, err := client.GetGroup(context.Background(), utils.GetAtomixRaftGroup())
-	if err != nil {
-		return nil, err
-	}
-
-	configs, err := group.GetMap(context.Background(), snapshotsName, session.WithTimeout(30*time.Second))
-	if err != nil {
-		return nil, err
-	}
-
-	return &atomixStore{
-		configs: configs,
-		closer:  configs,
-	}, nil
-}
-
-// NewLocalStore returns a new local device store
-func NewLocalStore() (Store, error) {
-	node, conn := startLocalNode()
-	configsName := primitive.Name{
-		Namespace: "local",
-		Name:      snapshotsName,
-	}
-	configs, err := _map.New(context.Background(), configsName, []*grpc.ClientConn{conn})
-	if err != nil {
-		return nil, err
-	}
-
-	return &atomixStore{
-		configs: configs,
-		closer:  utils.NewNodeCloser(node),
-	}, nil
-}
-
-// startLocalNode starts a single local node
-func startLocalNode() (*atomix.Node, *grpc.ClientConn) {
-	lis := bufconn.Listen(1024 * 1024)
-	node := local.NewNode(lis, registry.Registry)
-	_ = node.Start()
-
-	dialer := func(ctx context.Context, address string) (net.Conn, error) {
-		return lis.Dial()
-	}
-
-	conn, err := grpc.DialContext(context.Background(), snapshotsName, grpc.WithContextDialer(dialer), grpc.WithInsecure())
-	if err != nil {
-		panic("Failed to dial network configurations")
-	}
-	return node, conn
-}
-
-// Store is the interface for the snapshot store
-type Store interface {
-	io.Closer
-
-	// Store stores a snapshot
-	Store(snapshot *snapshottype.DeviceSnapshot) error
-
-	// Load loads a snapshot
-	Load(deviceID device.ID) (*snapshottype.DeviceSnapshot, error)
-}
-
-// atomixStore is the default implementation of the NetworkConfig store
-type atomixStore struct {
-	configs _map.Map
-	closer  io.Closer
-}
-
-func (s *atomixStore) Store(snapshot *snapshottype.DeviceSnapshot) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
-	defer cancel()
-
-	bytes, err := proto.Marshal(snapshot)
-	if err != nil {
-		return err
-	}
-
-	_, err = s.configs.Put(ctx, string(snapshot.DeviceID), bytes)
-	return err
-}
-
-func (s *atomixStore) Load(deviceID device.ID) (*snapshottype.DeviceSnapshot, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
-	defer cancel()
-
-	entry, err := s.configs.Get(ctx, string(deviceID))
-	if err != nil {
-		return nil, err
-	} else if entry == nil {
-		return nil, nil
-	}
-
-	snapshot := &snapshottype.DeviceSnapshot{}
-	if err := proto.Unmarshal(entry.Value, snapshot); err != nil {
-		return nil, err
-	}
-	return snapshot, nil
-}
-
-func (s *atomixStore) Close() error {
-	_ = s.configs.Close()
-	return s.closer.Close()
-}
