--- pkg/types/config/config.proto
@@ -1,77 +1,0 @@
-/*
-Copyright 2019-present Open Networking Foundation.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-syntax = "proto3";
-
-package onos.config.network;
-
-import "google/protobuf/timestamp.proto";
-import "gogoproto/gogo.proto";
-import "github.com/onosproject/onos-config/pkg/types/change/change.proto";
-
-// Status is the status of a configuration object
-enum Status {
-    // PENDING indicates the configuration is waiting to be applied
-    PENDING = 0;
-
-    // APPLYING indicates the configuration is in the process of being applied
-    APPLYING = 1;
-
-    // SUCCEEDED indicates the configuration was applied successfully
-    SUCCEEDED = 2;
-
-    // FAILED indicates the configuration failed
-    FAILED = 3;
-}
-
-// Reason is a failure reason
-enum Reason {
-    // ERROR indicates an error occurred
-    ERROR = 0;
-
-    // UNAVAILABLE indicates the device was unavailable
-    UNAVAILABLE = 1;
-}
-
-// NetworkConfig specifies the configuration for a network change
-message NetworkConfig {
-    // id is the ID of the configuration
-    string id = 1 [(gogoproto.customname) = "ID", (gogoproto.casttype) = "ID"];
-
-    // index is the index of the configuration
-    uint64 index = 2 [(gogoproto.casttype) = "Index"];
-
-    // revision is the configuration revision number
-    uint64 revision = 3 [(gogoproto.casttype) = "Revision"];
-
-    // status is the current status of the configuration
-    Status status = 4;
-
-    // reason indicates the reason the configuration failed
-    Reason reason = 5;
-
-    // message is an optional status message
-    string message = 6;
-
-    // created is the time at which the configuration was created
-    google.protobuf.Timestamp created = 7 [(gogoproto.stdtime) = true, (gogoproto.nullable) = false];
-
-    // updated is the time at which the configuration was last updated
-    google.protobuf.Timestamp updated = 8 [(gogoproto.stdtime) = true, (gogoproto.nullable) = false];
-
-    // changes is a set of changes to apply
-    repeated onos.config.change.Change changes = 9;
-}
