+++ pkg/types/device/change/types.proto
@@ -1,0 +1,119 @@
+/*
+Copyright 2019-present Open Networking Foundation.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+syntax = "proto3";
+
+package onos.config.device.change;
+
+import "google/protobuf/timestamp.proto";
+import "gogoproto/gogo.proto";
+
+// Status is the status of a configuration object
+enum Status {
+    // PENDING indicates the configuration is waiting to be applied
+    PENDING = 0;
+
+    // APPLYING indicates the configuration is in the process of being applied
+    APPLYING = 1;
+
+    // SUCCEEDED indicates the configuration was applied successfully
+    SUCCEEDED = 2;
+
+    // FAILED indicates the configuration failed
+    FAILED = 3;
+}
+
+// Reason is a failure reason
+enum Reason {
+    // ERROR indicates an error occurred
+    ERROR = 0;
+
+    // UNAVAILABLE indicates the device was unavailable
+    UNAVAILABLE = 1;
+}
+
+// Change represents a configuration change to a single device
+message Change {
+    // The identifier of the change
+    string id = 1 [(gogoproto.customname) = "ID", (gogoproto.casttype) = "ID"];
+
+    // The ID of the parent NetworkConfig
+    string config_id = 2 [(gogoproto.customname) = "ConfigID", (gogoproto.casttype) = "github.com/onosproject/onos-config/pkg/types.ID"];
+
+    // The configuration revision number
+    uint64 revision = 3 [(gogoproto.casttype) = "Revision"];
+
+    // The ID of the device to change
+    string device_id = 4 [(gogoproto.customname) = "DeviceID", (gogoproto.casttype) = "github.com/onosproject/onos-topo/pkg/northbound/device.ID"];
+
+    // The version of the device to change
+    string device_version = 5;
+
+    // The current status of the configuration
+    Status status = 6;
+
+    // The reason the change failed
+    Reason reason = 7;
+
+    // An optional status message
+    string message = 8;
+
+    // The time at which the configuration was created
+    google.protobuf.Timestamp created = 9 [(gogoproto.stdtime) = true, (gogoproto.nullable) = false];
+
+    // The time at which the configuration was last updated
+    google.protobuf.Timestamp updated = 10 [(gogoproto.stdtime) = true, (gogoproto.nullable) = false];
+
+    // The values to change
+    repeated Value values = 11;
+}
+
+// ValueType is the type for a value
+enum ValueType {
+    EMPTY = 0;
+    STRING = 1;
+    INT = 2;
+    UINT = 3;
+    BOOL = 4;
+    DECIMAL = 5;
+    FLOAT = 6;
+    BYTES = 7;
+    LEAFLIST_STRING = 8;
+    LEAFLIST_INT = 9;
+    LEAFLIST_UINT = 10;
+    LEAFLIST_BOOL = 11;
+    LEAFLIST_DECIMAL = 12;
+    LEAFLIST_FLOAT = 13;
+    LEAFLIST_BYTES = 14;
+}
+
+// Value is an individual Path/Value combination in a Change
+message Value {
+    // path is the path to change
+    string path = 1;
+
+    // value is the value
+    bytes value = 2;
+
+    // type is the value type
+    ValueType type = 3;
+
+    // type_opts is a set of type options
+    repeated int32 type_opts = 4;
+
+    // removed indicates whether this is a delete
+    bool removed = 5;
+}
