+++ pkg/types/device/snapshot/types.go
