+++ pkg/types/device/snapshot/types.proto
@@ -1,0 +1,89 @@
+/*
+Copyright 2019-present Open Networking Foundation.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+syntax = "proto3";
+
+package onos.config.device.snapshot;
+
+import "google/protobuf/timestamp.proto";
+import "gogoproto/gogo.proto";
+import "github.com/onosproject/onos-config/pkg/types/device/change/types.proto";
+
+// Status is the status of a configuration object
+enum Status {
+    // PENDING indicates the configuration is waiting to be applied
+    PENDING = 0;
+
+    // APPLYING indicates the configuration is in the process of being applied
+    APPLYING = 1;
+
+    // SUCCEEDED indicates the configuration was applied successfully
+    SUCCEEDED = 2;
+
+    // FAILED indicates the configuration failed
+    FAILED = 3;
+}
+
+// Reason is a failure reason
+enum Reason {
+    // ERROR indicates an error occurred
+    ERROR = 0;
+
+    // UNAVAILABLE indicates the device was unavailable
+    UNAVAILABLE = 1;
+}
+
+// DeviceSnapshot is a device snapshot request
+message DeviceSnapshot {
+    // the request identifier
+    string id = 1 [(gogoproto.customname) = "ID", (gogoproto.casttype) = "ID"];
+
+    // device_id is the device to which the snapshot applies
+    string device_id = 2 [(gogoproto.customname) = "DeviceID", (gogoproto.casttype) = "github.com/onosproject/onos-topo/pkg/northbound/device.ID"];
+
+    // revision is the request revision number
+    uint64 revision = 3 [(gogoproto.casttype) = "Revision"];
+
+    // status is the snapshot status
+    Status status = 4;
+
+    // reason is the snapshot failure reason
+    Reason reason = 5;
+
+    // timestamp is the time at which to take the snapshot
+    google.protobuf.Timestamp timestamp = 6 [(gogoproto.stdtime) = true, (gogoproto.nullable) = false];
+
+    // created is the time at which the configuration was created
+    google.protobuf.Timestamp created = 7 [(gogoproto.stdtime) = true, (gogoproto.nullable) = false];
+
+    // updated is the time at which the configuration was last updated
+    google.protobuf.Timestamp updated = 8 [(gogoproto.stdtime) = true, (gogoproto.nullable) = false];
+}
+
+// Snapshot is a snapshot of a single device
+message Snapshot {
+    // the request identifier
+    string id = 1 [(gogoproto.customname) = "ID", (gogoproto.casttype) = "ID"];
+
+    // device_id is the device to which the snapshot applies
+    string device_id = 2 [(gogoproto.customname) = "DeviceID", (gogoproto.casttype) = "github.com/onosproject/onos-topo/pkg/northbound/device.ID"];
+
+    // values is a list of values to set
+    repeated onos.config.device.change.Value values = 3;
+
+    // timestamp is the time at which to take the snapshot
+    google.protobuf.Timestamp timestamp = 4 [(gogoproto.stdtime) = true, (gogoproto.nullable) = false];
+}
