--- pkg/types/snapshot/snapshot.proto
@@ -1,38 +1,0 @@
-/*
-Copyright 2019-present Open Networking Foundation.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-syntax = "proto3";
-
-package onos.config.snapshot;
-
-import "google/protobuf/timestamp.proto";
-import "gogoproto/gogo.proto";
-import "github.com/onosproject/onos-config/pkg/types/change/change.proto";
-
-// DeviceSnapshot is a snapshot of a single device
-message DeviceSnapshot {
-    // the request identifier
-    string id = 1 [(gogoproto.customname) = "ID", (gogoproto.casttype) = "ID"];
-
-    // device_id is the device to which the snapshot applies
-    string device_id = 2 [(gogoproto.customname) = "DeviceID", (gogoproto.casttype) = "github.com/onosproject/onos-topo/pkg/northbound/device.ID"];
-
-    // values is a list of values to set
-    repeated onos.config.change.Value values = 3;
-
-    // timestamp is the time at which to take the snapshot
-    google.protobuf.Timestamp timestamp = 4 [(gogoproto.stdtime) = true, (gogoproto.nullable) = false];
-}
